+++ contrib/scripts/release.sh
@@ -10,7 +10,6 @@ args=(
     "-p nostr-indexeddb"
     "-p nostr-relay-builder"
     "-p nostr-relay-pool"
-    "-p nostr-zapper"
     "-p nwc"
     "-p nostr-connect"
     "-p nostr-sdk"