+++ crates/nostr-sdk/src/client/error.rs
@@ -9,8 +9,6 @@ use nostr::serde_json;
 use nostr_database::prelude::*;
 use nostr_relay_pool::__private::SharedStateError;
 use nostr_relay_pool::prelude::*;
-#[cfg(feature = "nip57")]
-use nostr_zapper::ZapperError;
 
 /// Client error
 #[derive(Debug)]
@@ -23,27 +21,15 @@ pub enum Error {
     Database(DatabaseError),
     /// Signer error
     Signer(SignerError),
-    /// Zapper error
-    #[cfg(feature = "nip57")]
-    Zapper(ZapperError),
     /// [`EventBuilder`] error
     EventBuilder(builder::Error),
     /// Json error
     Json(serde_json::Error),
     /// Shared state error
     SharedState(SharedStateError),
-    /// NIP57 error
-    #[cfg(feature = "nip57")]
-    NIP57(nip57::Error),
-    /// LNURL Pay
-    #[cfg(feature = "nip57")]
-    LnUrlPay(lnurl_pay::Error),
     /// NIP59
     #[cfg(feature = "nip59")]
     NIP59(nip59::Error),
-    /// Zapper not configured
-    #[cfg(feature = "nip57")]
-    ZapperNotConfigured,
     /// Event not found
     EventNotFound(EventId),
     /// Impossible to zap
@@ -65,21 +51,11 @@ impl fmt::Display for Error {
             Self::RelayPool(e) => write!(f, "{e}"),
             Self::Database(e) => write!(f, "{e}"),
             Self::Signer(e) => write!(f, "{e}"),
-            #[cfg(feature = "nip57")]
-            Self::Zapper(e) => write!(f, "{e}"),
             Self::EventBuilder(e) => write!(f, "{e}"),
             Self::Json(e) => write!(f, "{e}"),
             Self::SharedState(e) => write!(f, "{e}"),
-            #[cfg(feature = "nip57")]
-            Self::NIP57(e) => write!(f, "{e}"),
-            #[cfg(feature = "nip57")]
-            Self::LnUrlPay(e) => write!(f, "{e}"),
             #[cfg(feature = "nip59")]
             Self::NIP59(e) => write!(f, "{e}"),
-            #[cfg(feature = "nip57")]
-            Self::ZapperNotConfigured => {
-                write!(f, "zapper not configured")
-            }
             Self::EventNotFound(id) => {
                 write!(f, "event not found: {id}")
             }
@@ -119,13 +95,6 @@ impl From<SignerError> for Error {
     }
 }
 
-#[cfg(feature = "nip57")]
-impl From<ZapperError> for Error {
-    fn from(e: ZapperError) -> Self {
-        Self::Zapper(e)
-    }
-}
-
 impl From<builder::Error> for Error {
     fn from(e: builder::Error) -> Self {
         Self::EventBuilder(e)
@@ -141,20 +110,6 @@ impl From<serde_json::Error> for Error {
 impl From<SharedStateError> for Error {
     fn from(e: SharedStateError) -> Self {
         Self::SharedState(e)
-    }
-}
-
-#[cfg(feature = "nip57")]
-impl From<nip57::Error> for Error {
-    fn from(e: nip57::Error) -> Self {
-        Self::NIP57(e)
-    }
-}
-
-#[cfg(feature = "nip57")]
-impl From<lnurl_pay::Error> for Error {
-    fn from(e: lnurl_pay::Error) -> Self {
-        Self::LnUrlPay(e)
     }
 }