+++ crates/nostr-sdk/src/prelude.rs
@@ -12,8 +12,6 @@
 pub use nostr::prelude::*;
 pub use nostr_database::prelude::*;
 pub use nostr_relay_pool::prelude::*;
-#[cfg(feature = "nip57")]
-pub use nostr_zapper::prelude::*;
 
 // Internal modules
 pub use crate::client::*;